update ramp code
