+++ examples/lifelong_learning/cityscapes/robo_detection/x86.Dockerfile
@@ -17,14 +17,18 @@ COPY ./requirements.txt /home/requirements.txt
 
 ENV PYTHONHTTPSVERIFY "0"
 ENV PROMPT_DIRTRIM "1"
+ENV BIG_MODEL_IP "http://94.74.91.114"
+ENV BIG_MODEL_PORT "30001"
 ENV PYTHONPATH "/home/lib"
 ENV HOLD_TIME 1
-ENV PIP "https://pypi.tuna.tsinghua.edu.cn/simple"
-RUN pip3 install -i $PIP --no-cache-dir -r /home/requirements.txt
-RUN pip3 install -i $PIP --no-cache-dir -r /home/requirements-sdk.txt
+
+RUN pip3 install --no-cache-dir -r /home/requirements.txt
+RUN pip3 install --no-cache-dir -r /home/requirements-sdk.txt
 
 COPY ./run.sh /usr/local/bin/run.sh
 RUN chmod 755 /usr/local/bin/run.sh
+COPY ./ramp_run.sh /usr/local/bin/ramp_run.sh
+RUN chmod 755 /usr/local/bin/ramp_run.sh
 
 WORKDIR /home/lib
 COPY ./robosdk /home/lib/robosdk
@@ -47,11 +51,6 @@ RUN rm -rf /var/lib/apt/lists/*
 ENV http_proxy ""
 ENV https_proxy ""
 ENV no_proxy ""
+# entrypoint
 
-RUN source /opt/ros/$ROS_DISTRO/setup.bash
-RUN source /home/deep_msgs/devel/setup.bash
-
-WORKDIR /home/lib/ramp_detection
-
-# ENTRYPOINT ["bash"]
-ENTRYPOINT ["python3"]
+ENTRYPOINT ["bash"]
